Change pqdm download tests to integration tests
